feat(camera client): split a cam-client service from app
Together, the cam-client and the host's cam-daemon constitute a new camera block. This camera block is independent from the app container, and vice-versa. They can be spun up independently. The daemon is independent from the cam-client, but the cam-client depends on the daemon.

A run script is provided, which allows for the easy spinning up of the camera block, the app, or both.

Finally, gave camera-daemon its own venv instead of relying on user-global package installations.
