+++ backend/.env.example
@@ -8,5 +8,8 @@ DOCKER=1
 # Set the logical name of the camera to use (must be created in camera-daemon first)
 CAMERA_NAME=camera0
 
+# Set the device path for the camera (e.g., /dev/video0)
+CAMERA_DEVICE=/dev/video0
+
 # Host device group id
 HOST_DEV_GID=
