+++ backend/devices/camera/client/healthcheck.py
@@ -1,0 +1,2 @@
+import urllib.request
+urllib.request.urlopen("http://localhost:9000/health")
