+++ backend/devices/camera/host/camera_daemon.py
