+++ backend/devices/camera/host/camera_paths.py
