+++ backend/devices/camera/host/logical_camera.py
