+++ backend/devices/camera/host/requirements.txt
@@ -1,0 +1,2 @@
+imageio
+opencv-python-headless
