+++ backend/devices/camera/host/selinux/aquadash_camera.te
@@ -1,0 +1,10 @@
+module aquadash_camera 1.0;
+
+require {
+    type container_t;
+    class unix_stream_socket connectto;
+    type container_file_t;
+}
+
+# Allow containers to connect to socket files labeled container_file_t
+allow container_t container_file_t:unix_stream_socket connectto;
