+++ backend/devices/camera/scripts/uninstall.sh
@@ -12,15 +12,15 @@ SERVICES=("camera-daemon.service")
 
 echo "==> Uninstalling Aquadash camera service"
 
-# --- Stop, disable, and remove systemd services ---
+# --- Stop and disable systemd services ---
 for svc in "${SERVICES[@]}"; do
     if systemctl is-active --quiet "$svc" 2>/dev/null; then
         echo "  Stopping $svc"
-        systemctl stop "$svc"
+        systemctl stop "$svc" || true
     fi
     if systemctl is-enabled --quiet "$svc" 2>/dev/null; then
         echo "  Disabling $svc"
-        systemctl disable "$svc"
+        systemctl disable "$svc" || true
     fi
     if [ -f "$SYSTEMD_DIR/$svc" ]; then
         echo "  Removing $SYSTEMD_DIR/$svc"
@@ -30,37 +30,47 @@ done
 
 systemctl daemon-reload
 
+# --- Kill any lingering processes ---
+echo "  Killing any remaining Python processes from $INSTALL_DIR"
+pkill -f "$INSTALL_DIR" >/dev/null 2>&1 || true
+sleep 1  # allow processes to exit
+
+# --- Remove /run/camera and SELinux contexts ---
+echo "  Removing /run/camera directory and SELinux policy if present..."
+if [ -d /run/camera ]; then
+    rm -rf /run/camera
+fi
+
+if command -v selinuxenabled >/dev/null 2>&1 && selinuxenabled; then
+    semodule -r aquadash_camera >/dev/null 2>&1 || true
+    semanage fcontext -d '/run/camera(/.*)?' >/dev/null 2>&1 || true
+    semanage fcontext -d '/run/camera' >/dev/null 2>&1 || true
+    restorecon -Rv /run/camera >/dev/null 2>&1 || true
+fi
+
 # --- Remove installed files ---
 echo "  Removing camera-ctl wrapper from $BIN_DIR"
 rm -f "$BIN_DIR/camera-ctl"
 
-echo "  Removing camera daemon from $INSTALL_DIR"
+echo "  Removing camera daemon directory $INSTALL_DIR"
 rm -rf "$INSTALL_DIR"
 
 # --- Remove user ---
-if id "$USER_NAME" > /dev/null 2>&1; then
+if id "$USER_NAME" >/dev/null 2>&1; then
     echo "  Removing user: $USER_NAME"
-    userdel "$USER_NAME"
+    pkill -u "$USER_NAME" >/dev/null 2>&1 || true
+    userdel "$USER_NAME" || true
 fi
 
 # --- Remove group (only if no remaining members) ---
-if getent group "$GROUP_NAME" > /dev/null 2>&1; then
+if getent group "$GROUP_NAME" >/dev/null 2>&1; then
     members="$(getent group "$GROUP_NAME" | cut -d: -f4)"
     if [ -z "$members" ]; then
         echo "  Removing group: $GROUP_NAME"
-        groupdel "$GROUP_NAME"
+        groupdel "$GROUP_NAME" || true
     else
         echo "  Skipping group removal: $GROUP_NAME still has members ($members)"
     fi
-fi
-
-# --- Remove SELinux policy (if installed) ---
-if command -v selinuxenabled >/dev/null 2>&1 && selinuxenabled; then
-    echo "  Removing SELinux camera policy..."
-
-    semodule -r aquadash_camera >/dev/null 2>&1 || true
-    semanage fcontext -d '/run/camera(/.*)?' >/dev/null 2>&1 || true
-    restorecon -Rv /run/camera >/dev/null 2>&1 || true
 fi
 
 echo ""
