+++ backend/docker-compose.yml
@@ -10,7 +10,7 @@ services:
       POSTGRES_DB: aquapoly
 
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U postgres"]
+      test: ["CMD", "pg_isready", "-U", "postgres"]
       interval: 5s
       timeout: 5s
       retries: 5
@@ -54,6 +54,33 @@ services:
 
     ports:
       - "8000:8000"
+
+    networks:
+      - backend
+  
+  cam_client:
+    build:
+      context: ./devices/camera/client
+      dockerfile: cam_client.Dockerfile
+
+    container_name: cam-client
+
+    environment:
+      CAMERA_NAME: ${CAMERA_NAME:-camera0}
+
+    healthcheck:
+      test: ["CMD", "python", "healthcheck.py"]
+      interval: 10s
+      timeout: 3s
+      retries: 3
+      start_period: 5s
+
+    restart: always
+
+    volumes:
+      - /run/camera:/run/camera:z
+
+    user: "1000:${HOST_DEV_GID}"
 
     networks:
       - backend
